move validation functions into a separate module
